--- Procfile
@@ -1,1 +1,0 @@
-web: gunicorn unidays.checkout_api:app
