+++ unidays/unidays.py
@@ -1,4 +1,4 @@
-from unidays.utils import errors, classInjectionMap, itemValidatorMap
+from utils import errors, classInjectionMap, itemValidatorMap
 
 class ErrorLogger:
     def __init__(self, errorMessages):
